+++ NEWS.md
@@ -4,6 +4,7 @@
 - MESS3D now calculated using modEvA::MESS() instead of dismo::mess().
 - Removed dependency on `raster` and `rgdal`. Raster visualizations now done with `terra` and include optional graticule plotting.
 - Added function to generate `SpatRaster` template with input `SpatVector` points centered on template cells.
+- Fixed messy axis labels when graticules are plotted for rasters.
 
 # voluModel 0.1.9